Fix fast refresh compatibility for filters
